requirements.txt
# Core dependencies with compatible versions
matplotlib>=3.7.0,<3.11.0
numpy>=1.23.0,<2.0.0
pandas>=2.0.0,<2.4.0
PyYAML>=6.0.0
scikit-learn>=1.3.0,<1.7.0
scipy>=1.10.0,<1.16.0
seaborn>=0.12.0,<0.14.0
shap>=0.40.0,<0.50.0
statsmodels>=0.14.0,<0.15.0
torch>=2.0.0,<2.8.0
torch-geometric>=2.4.0,<2.7.0
tqdm>=4.60.0
wandb>=0.15.0,<0.22.0
xgboost>=2.0.0,<3.1.0

# Additional dependencies for video_inference.py
opencv-python>=4.8.0,<4.11.0
easyocr>=1.6.0,<1.8.0
ultralytics>=8.0.0,<8.4.0
python-dotenv>=1.0.0
streamlink>=6.0.0,<6.12.0
requests>=2.28.0,<2.33.0
imageio>=2.30.0,<2.37.0
Pillow>=10.0.0,<10.5.0

# FastAPI backend dependencies
fastapi>=0.104.0,<0.115.0
uvicorn[standard]>=0.24.0,<0.32.0
python-multipart>=0.0.6,<0.0.10
websockets>=12.0,<13.0

# Ultralytics dependencies that might be auto-installed
lap>=0.5.12
